+++ proxywasm/hostcall.go
@@ -15,7 +15,11 @@
 package proxywasm
 
 import (
+	"fmt"
+	"math"
+
 	"github.com/tetratelabs/proxy-wasm-go-sdk/proxywasm/internal"
+	"github.com/tetratelabs/proxy-wasm-go-sdk/proxywasm/types"
 )
 
 // GetVMConfiguration is used for retrieving configurations ginve in the "vm_config.configuration" field.
@@ -555,4 +559,260 @@ func CallForeignFunction(funcName string, param []byte) (ret []byte, err error) 
 	default:
 		return nil, internal.StatusToError(st)
 	}
+}
+
+// LogTracef emit a message as a log with Trace log level.
+func LogTrace(msg string) {
+	internal.ProxyLog(internal.LogLevelTrace, internal.StringBytePtr(msg), len(msg))
+}
+
+// LogTracef formats according to a format specifier and emit as a log with Trace log level.
+func LogTracef(format string, args ...interface{}) {
+	msg := fmt.Sprintf(format, args...)
+	internal.ProxyLog(internal.LogLevelTrace, internal.StringBytePtr(msg), len(msg))
+}
+
+// LogTracef emit a message as a log with Debug log level.
+func LogDebug(msg string) {
+	internal.ProxyLog(internal.LogLevelDebug, internal.StringBytePtr(msg), len(msg))
+}
+
+// LogDebugf formats according to a format specifier and emit as a log with Debug log level.
+func LogDebugf(format string, args ...interface{}) {
+	msg := fmt.Sprintf(format, args...)
+	internal.ProxyLog(internal.LogLevelDebug, internal.StringBytePtr(msg), len(msg))
+}
+
+// LogTracef emit a message as a log with Info log level.
+func LogInfo(msg string) {
+	internal.ProxyLog(internal.LogLevelInfo, internal.StringBytePtr(msg), len(msg))
+}
+
+// LogInfof formats according to a format specifier and emit as a log with Info log level.
+func LogInfof(format string, args ...interface{}) {
+	msg := fmt.Sprintf(format, args...)
+	internal.ProxyLog(internal.LogLevelInfo, internal.StringBytePtr(msg), len(msg))
+}
+
+// LogTracef emit a message as a log with Warn log level.
+func LogWarn(msg string) {
+	internal.ProxyLog(internal.LogLevelWarn, internal.StringBytePtr(msg), len(msg))
+}
+
+// LogWarnf formats according to a format specifier and emit as a log with Warn log level.
+func LogWarnf(format string, args ...interface{}) {
+	msg := fmt.Sprintf(format, args...)
+	internal.ProxyLog(internal.LogLevelWarn, internal.StringBytePtr(msg), len(msg))
+}
+
+// LogTracef emit a message as a log with Error log level.
+func LogError(msg string) {
+	internal.ProxyLog(internal.LogLevelError, internal.StringBytePtr(msg), len(msg))
+}
+
+// LogErrorf formats according to a format specifier and emit as a log with Error log level.
+func LogErrorf(format string, args ...interface{}) {
+	msg := fmt.Sprintf(format, args...)
+	internal.ProxyLog(internal.LogLevelError, internal.StringBytePtr(msg), len(msg))
+}
+
+// LogTracef emit a message as a log with Critical log level.
+func LogCritical(msg string) {
+	internal.ProxyLog(internal.LogLevelCritical, internal.StringBytePtr(msg), len(msg))
+}
+
+// LogCriticalf formats according to a format specifier and emit as a log with Critical log level.
+func LogCriticalf(format string, args ...interface{}) {
+	msg := fmt.Sprintf(format, args...)
+	internal.ProxyLog(internal.LogLevelCritical, internal.StringBytePtr(msg), len(msg))
+}
+
+type (
+	// MetricCounter represents a counter metric.
+	// Use DefineCounterMetric for initialization.
+	MetricCounter uint32
+	// MetricGauge represents a gauge metric.
+	// Use DefineGaugeMetric for initialization.
+	MetricGauge uint32
+	// MetricHistogram represents a histogram metric.
+	// Use DefineHistogramMetric for initialization.
+	MetricHistogram uint32
+)
+
+// DefineCounterMetric returnes MetricCounter for a name.
+func DefineCounterMetric(name string) MetricCounter {
+	var id uint32
+	ptr := internal.StringBytePtr(name)
+	st := internal.ProxyDefineMetric(internal.MetricTypeCounter, ptr, len(name), &id)
+	if err := internal.StatusToError(st); err != nil {
+		panic(fmt.Sprintf("define metric of name %s: %v", name, internal.StatusToError(st)))
+	}
+	return MetricCounter(id)
+}
+
+// Value returnes the current value for this counter.
+func (m MetricCounter) Value() uint64 {
+	var val uint64
+	st := internal.ProxyGetMetric(uint32(m), &val)
+	if err := internal.StatusToError(st); err != nil {
+		panic(fmt.Sprintf("get metric of  %d: %v", uint32(m), internal.StatusToError(st)))
+	}
+	return val
+}
+
+// Increment increments the current value by a offset for this counter.
+func (m MetricCounter) Increment(offset uint64) {
+	if err := internal.StatusToError(internal.ProxyIncrementMetric(uint32(m), int64(offset))); err != nil {
+		panic(fmt.Sprintf("increment %d by %d: %v", uint32(m), offset, err))
+	}
+}
+
+// DefineCounterMetric returnes MetricGauge for a name.
+func DefineGaugeMetric(name string) MetricGauge {
+	var id uint32
+	ptr := internal.StringBytePtr(name)
+	st := internal.ProxyDefineMetric(internal.MetricTypeGauge, ptr, len(name), &id)
+	if err := internal.StatusToError(st); err != nil {
+		panic(fmt.Sprintf("error define metric of name %s: %v", name, internal.StatusToError(st)))
+	}
+	return MetricGauge(id)
+}
+
+// Value returnes the current value for this gauge.
+func (m MetricGauge) Value() int64 {
+	var val uint64
+	if err := internal.StatusToError(internal.ProxyGetMetric(uint32(m), &val)); err != nil {
+		panic(fmt.Sprintf("get metric of  %d: %v", uint32(m), err))
+	}
+	return int64(val)
+}
+
+// Add adds a offset to the current value for this counter.
+func (m MetricGauge) Add(offset int64) {
+	if err := internal.StatusToError(internal.ProxyIncrementMetric(uint32(m), offset)); err != nil {
+		panic(fmt.Sprintf("error adding %d by %d: %v", uint32(m), offset, err))
+	}
+}
+
+// DefineHistogramMetric returnes MetricHistogram for a name.
+func DefineHistogramMetric(name string) MetricHistogram {
+	var id uint32
+	ptr := internal.StringBytePtr(name)
+	st := internal.ProxyDefineMetric(internal.MetricTypeHistogram, ptr, len(name), &id)
+	if err := internal.StatusToError(st); err != nil {
+		panic(fmt.Sprintf("error define metric of name %s: %v", name, internal.StatusToError(st)))
+	}
+	return MetricHistogram(id)
+}
+
+// Value returnes the current value for this histogram.
+func (m MetricHistogram) Value() uint64 {
+	var val uint64
+	st := internal.ProxyGetMetric(uint32(m), &val)
+	if err := internal.StatusToError(st); err != nil {
+		panic(fmt.Sprintf("get metric of  %d: %v", uint32(m), internal.StatusToError(st)))
+	}
+	return val
+}
+
+// Record records a value for this histogram.
+func (m MetricHistogram) Record(value uint64) {
+	if err := internal.StatusToError(internal.ProxyRecordMetric(uint32(m), value)); err != nil {
+		panic(fmt.Sprintf("error adding %d: %v", uint32(m), err))
+	}
+}
+
+func setMap(mapType internal.MapType, headers [][2]string) error {
+	shs := internal.SerializeMap(headers)
+	hp := &shs[0]
+	hl := len(shs)
+	return internal.StatusToError(internal.ProxySetHeaderMapPairs(mapType, hp, hl))
+}
+
+func getMapValue(mapType internal.MapType, key string) (string, error) {
+	var rvs int
+	var raw *byte
+	if st := internal.ProxyGetHeaderMapValue(
+		mapType, internal.StringBytePtr(key), len(key), &raw, &rvs,
+	); st != internal.StatusOK {
+		return "", internal.StatusToError(st)
+	}
+
+	ret := internal.RawBytePtrToString(raw, rvs)
+	return ret, nil
+}
+
+func removeMapValue(mapType internal.MapType, key string) error {
+	return internal.StatusToError(
+		internal.ProxyRemoveHeaderMapValue(mapType, internal.StringBytePtr(key), len(key)),
+	)
+}
+
+func replaceMapValue(mapType internal.MapType, key, value string) error {
+	return internal.StatusToError(
+		internal.ProxyReplaceHeaderMapValue(
+			mapType, internal.StringBytePtr(key), len(key), internal.StringBytePtr(value), len(value),
+		),
+	)
+}
+
+func addMapValue(mapType internal.MapType, key, value string) error {
+	return internal.StatusToError(
+		internal.ProxyAddHeaderMapValue(
+			mapType, internal.StringBytePtr(key), len(key), internal.StringBytePtr(value), len(value),
+		),
+	)
+}
+
+func getMap(mapType internal.MapType) ([][2]string, error) {
+	var rvs int
+	var raw *byte
+
+	st := internal.ProxyGetHeaderMapPairs(mapType, &raw, &rvs)
+	if st != internal.StatusOK {
+		return nil, internal.StatusToError(st)
+	}
+
+	bs := internal.RawBytePtrToByteSlice(raw, rvs)
+	return internal.DeserializeMap(bs), nil
+}
+
+func getBuffer(bufType internal.BufferType, start, maxSize int) ([]byte, error) {
+	var retData *byte
+	var retSize int
+	switch st := internal.ProxyGetBufferBytes(bufType, start, maxSize, &retData, &retSize); st {
+	case internal.StatusOK:
+		if retData == nil {
+			return nil, types.ErrorStatusNotFound
+		}
+		return internal.RawBytePtrToByteSlice(retData, retSize), nil
+	default:
+		return nil, internal.StatusToError(st)
+	}
+}
+
+func appendToBuffer(bufType internal.BufferType, buffer []byte) error {
+	var bufferData *byte
+	if len(buffer) != 0 {
+		bufferData = &buffer[0]
+	}
+	return internal.StatusToError(internal.ProxySetBufferBytes(bufType, math.MaxInt32, 0, bufferData, len(buffer)))
+}
+
+func prependToBuffer(bufType internal.BufferType, buffer []byte) error {
+	var bufferData *byte
+	if len(buffer) != 0 {
+		bufferData = &buffer[0]
+	}
+	return internal.StatusToError(internal.ProxySetBufferBytes(bufType, 0, 0, bufferData, len(buffer)))
+}
+
+func replaceBuffer(bufType internal.BufferType, buffer []byte) error {
+	var bufferData *byte
+	if len(buffer) != 0 {
+		bufferData = &buffer[0]
+	}
+	return internal.StatusToError(
+		internal.ProxySetBufferBytes(bufType, 0, math.MaxInt32, bufferData, len(buffer)),
+	)
 }
