+++ proxywasm/proxytest/README.md
@@ -6,7 +6,7 @@ With `proxytest` build tag, you can test your extension with the official comman
 go test -tags=proxytest ./... 
 ```
 
-This framework simulates the expected behavior of Envoyproxy, and you can test your extensions without running Envoy. 
+This framework emulates the expected behavior of Envoyproxy, and you can test your extensions without running Envoy.
 For detail, see `examples/*/main_test.go`.
